Generazione progetto PlatformIO (Compile OK) Riorganizzazione file e cancellazione file inutili Correzione sorgenti per errori e warning in compilazione

+++ lib/ButtonSet/ButtonAdv.cpp
@@ -182,7 +182,7 @@ ButtonAdv::_check(uint8_t newi)
             // is only used to spare computing time, skipping the check of the whole condition.
             if ((TlastPress != 0) || now >= TstartPress + times100(repeatDelay)) {
                 // is it time for a repeat keypress call?
-                if ((now - TlastPress) >= times10(repeatRate)) {
+                if ((now - TlastPress) >= (unsigned long)times10(repeatRate)) {
                     TlastPress = now;
                     _OnPress(this);
                     //now = millis();     // callback may have taken some time
